refactor(fleet-console): move locale date formatting into the SDK
Relative and absolute date formatting is pure and reads no host state, so the
SDK owns it and core folds its path. A plugin that picks its own Intl options
makes the same instant read differently from one surface to the next inside a
single console.
